requirements.txt
Flask
google-generativeai
twilio
gunicorn
